requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.19.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.11
multidict==6.0.4
python-dotenv==1.0.0
pytz==2023.3
yarl==1.9.2
